refactor: 노마드 리팩터링 1차 완료
